Moved single drafts action into messages module.
